增加基础功能

+++ src/Controller/DeleteStoreController.php
@@ -7,8 +7,9 @@ use Flarum\Api\Controller\AbstractListController;
 use Flarum\Http\RequestUtil;
 use Flarum\Http\UrlGenerator;
 use Flarum\Locale\Translator;
+use Flarum\User\Exception\PermissionDeniedException;
 use Flarum\User\UserRepository;
-use Mattoid\Store\Model\StoreCommodityModel;
+use Mattoid\Store\Model\StoreGoodsModel;
 use Mattoid\Store\Model\StoreModel;
 use Mattoid\Store\Serializer\StoreSerializer;
 use Psr\Http\Message\ServerRequestInterface;
@@ -32,6 +33,10 @@ class DeleteStoreController extends AbstractListController
     protected function data(ServerRequestInterface $request, Document $document) {
         $actor = RequestUtil::getActor($request);
         $parseBody = $request->getParsedBody();
+
+        if (!$actor->can('mattoid-store.group-moderate')) {
+            throw new PermissionDeniedException();
+        }
 
         $result = StoreModel::query()->where('id', $parseBody['id'])->delete();
